Conflicts:
	DataAccess/Entities/UserAndPhotographer.cs
	DataAccessAbstraction/Entities/IUserAndPhotographer.cs
	Logger/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
	PhotoGBG.v11.suo
	Storage/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
	Utilities/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache

+++ DataAccessAbstraction/Entities/ITag.cs
@@ -8,7 +8,7 @@ namespace DataAccessAbstraction.Entities
 {
     public interface ITag
     {
-        long Id { get; set; }
+        int Id { get; set; }
         string Name { get; set; }
         long AddedBy { get; set; } //id field from user db
     }
